Moved Activity to giphy package
