+++ app/src/main/java/io/jeffchang/challenge/di/module/MainActivityModule.kt
@@ -3,7 +3,7 @@ package io.jeffchang.challenge.di.module
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
 import io.jeffchang.giphy.GiphyModule
-import io.jeffchang.challenge.MainActivity
+import io.jeffchang.giphy.GiphySearchActivity
 
 @Module
 abstract class MainActivityModule {
@@ -11,6 +11,6 @@ abstract class MainActivityModule {
     @ContributesAndroidInjector(modules = [
         GiphyModule::class
     ])
-    abstract fun contributeMainActivity(): MainActivity
+    abstract fun contributeMainActivity(): GiphySearchActivity
 
 }
