+++ app/src/main/java/io/jeffchang/giphy/GiphySearchActivity.kt
@@ -1,14 +1,21 @@
-package io.jeffchang.challenge
+package io.jeffchang.giphy
 
 import android.os.Bundle
+import android.view.Menu
 import androidx.navigation.Navigation.findNavController
 import dagger.android.support.DaggerAppCompatActivity
+import io.jeffchang.challenge.R
 
-class MainActivity : DaggerAppCompatActivity() {
+class GiphySearchActivity : DaggerAppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
+        setContentView(R.layout.activity_giphy_search)
+    }
+
+    override fun onCreateOptionsMenu(menu: Menu): Boolean {
+        menuInflater.inflate(R.menu.giphy_search_options_menu, menu)
+        return true
     }
 
     override fun onSupportNavigateUp() =
