+++ giphy/src/main/res/navigation/nav_giphy_search.xml
